changed to object protocol, everything works besides observers and logout on multiple windows

+++ model/src/main/java/transportAgency/model/Seat.java
@@ -1,3 +1,3 @@
 package transportAgency.model;
 
-public record Seat(Long seatNo, String clientName) {}
+public record Seat(Integer seatNo, String clientName) {}
